container: load ticket during start
Use an environment variable, namely
TICKET_PATH, to allow users to define
the full path for the volume mount point
and the ticket name, so that the file
can be used in the CMD instruction,
and the containerized imageio
can load the ticket from start.

The volume containing the image should be
mounted at runtime in the path specified
in the ticket, and have the right permissions.

Example run command:

TICKET_PATH=/ticket/ticket.json podman run \
    --env-host --rm \
    --privileged \
    -it -p 8080:80 \
    --volume test:/ticket \
    --volume /path/to/image:/images:Z \
    localhost/ovirt-imageio:latest

Signed-off-by: Albert Esteve <aesteve@redhat.com>